CloudFunctions/gcs2gcs/requirements.txt
cloudevents
google-cloud-storage
functions-framework==3.*